+++ src/integration_tests/test_kprint.py
@@ -1,6 +1,6 @@
 from pyk.kast import KApply, KToken
 from pyk.ktool import KompileBackend
-from pyk.prelude import intToken
+from pyk.prelude.kint import intToken
 
 from .kprove_test import KProveTest
