+++ src/integration_tests/test_parse_token.py
@@ -3,7 +3,7 @@ from typing import Iterable, Tuple
 from pyk.kast import KApply, KAtt, KInner, KToken, KVariable
 from pyk.ktool import KompileBackend
 from pyk.ktool.kit import parse_token_rule_syntax
-from pyk.prelude import intToken
+from pyk.prelude.kint import intToken
 from pyk.utils import FrozenDict
 
 from .kprove_test import KProveTest
