+++ src/pyk/cterm.py
@@ -16,7 +16,7 @@ from .kastManip import (
     split_config_and_constraints,
     substitute,
 )
-from .prelude import mlAnd, mlImplies, mlTop
+from .prelude.ml import mlAnd, mlImplies, mlTop
 from .prelude.sorts import GENERATED_TOP_CELL
 from .utils import unique
