+++ src/pyk/kastManip.py
@@ -20,8 +20,9 @@ from .kast import (
     collect,
     top_down,
 )
-from .prelude import DOTS, Labels, mlAnd, mlBottom, mlEqualsTrue, mlImplies, mlOr, mlTop
+from .prelude import DOTS, Labels
 from .prelude.kbool import FALSE, TRUE, andBool, impliesBool, notBool, orBool
+from .prelude.ml import mlAnd, mlBottom, mlEqualsTrue, mlImplies, mlOr, mlTop
 from .prelude.sorts import GENERATED_TOP_CELL
 from .utils import find_common_items, hash_str
