+++ src/tests/test_kcfg.py
@@ -4,8 +4,9 @@ from unittest import TestCase
 from pyk.cterm import CTerm
 from pyk.kast import KApply, KInner, KVariable
 from pyk.kcfg import KCFG
-from pyk.prelude import mlEquals, token
+from pyk.prelude import token
 from pyk.prelude.kbool import TRUE
+from pyk.prelude.ml import mlEquals
 from pyk.utils import shorten_hash
 
 from .mock_kprint import MockKPrint
